thermal: Update THERMAL_MAX_TRIPS to 48
Also changed mask field from int to u64 to support the larger trip point
number.

Bug 2125602

Reviewed-on: http://git-master/r/327982
(cherry picked from commit 8236da80e83a6ede9b1f70741194024c15d975dc)
(wni, fix conflicts, some routines were moved from thermal_core.c to
thermal_sys.c in K4.14)
Reviewed-on: https://git-master.nvidia.com/r/1736760
(cherry picked from commit 0572b6c909ade414a6d5bfe35bba9de9a0519f0f)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2410491
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/thermal/thermal_core.c
@@ -1025,7 +1025,7 @@ void print_bind_err_msg(struct thermal_zone_device *tz,
 		tz->type, cdev->type, ret);
 }
 
-static void __bind(struct thermal_zone_device *tz, int mask,
+static void __bind(struct thermal_zone_device *tz, u64 mask,
 		   struct thermal_cooling_device *cdev,
 		   unsigned long *limits,
 		   unsigned int weight)
@@ -1033,7 +1033,7 @@ static void __bind(struct thermal_zone_device *tz, int mask,
 	int i, ret;
 
 	for (i = 0; i < tz->trips; i++) {
-		if (mask & (1 << i)) {
+		if (mask & (1ULL << i)) {
 			unsigned long upper, lower;
 
 			upper = THERMAL_NO_LIMIT;
@@ -1261,13 +1261,13 @@ devm_thermal_of_cooling_device_register(struct device *dev,
 }
 EXPORT_SYMBOL_GPL(devm_thermal_of_cooling_device_register);
 
-static void __unbind(struct thermal_zone_device *tz, int mask,
+static void __unbind(struct thermal_zone_device *tz, u64 mask,
 		     struct thermal_cooling_device *cdev)
 {
 	int i;
 
 	for (i = 0; i < tz->trips; i++)
-		if (mask & (1 << i))
+		if (mask & (1ULL << i))
 			thermal_zone_unbind_cooling_device(tz, i, cdev);
 }
 
@@ -1392,7 +1392,7 @@ exit:
  * IS_ERR*() helpers.
  */
 struct thermal_zone_device *
-thermal_zone_device_register(const char *type, int trips, int mask,
+thermal_zone_device_register(const char *type, int trips, u64 mask,
 			     void *devdata, struct thermal_zone_device_ops *ops,
 			     struct thermal_zone_params *tzp, int passive_delay,
 			     int polling_delay)
