+++ drivers/thermal/thermal_core.h
@@ -135,7 +135,7 @@ int thermal_build_list_of_policies(char *buf);
 void thermal_zone_set_trips(struct thermal_zone_device *tz);
 
 /* sysfs I/F */
-int thermal_zone_create_device_groups(struct thermal_zone_device *, int);
+int thermal_zone_create_device_groups(struct thermal_zone_device *, u64);
 void thermal_zone_destroy_device_groups(struct thermal_zone_device *);
 void thermal_cooling_device_setup_sysfs(struct thermal_cooling_device *);
 void thermal_cooling_device_destroy_sysfs(struct thermal_cooling_device *cdev);
